% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsFPOP.R
\name{MsFPOP}
\alias{MsFPOP}
\title{MsFPOP}
\usage{
MsFPOP(
  y,
  beta,
  alpha,
  sampling_method = "rand",
  sampling_method_parameter = 1,
  wt
)
}
\arguments{
\item{y}{A vector of observations.}

\item{beta}{A constant used in the the penalty.}

\item{alpha}{A constant used in the the penalty.
(recommended: gamma+beta*log(length(y)))}

\item{sampling_method}{A sampling method for changepoint candidates ("rand, "all").}

\item{sampling_method_parameter}{A parameter used by the sampling method.}

\item{wt}{A vector of weights associated to the observations.}
}
\value{
A list that includes the estimated position of the changepoints.
}
\description{
MsFPOP is an exact segmentation algorithm that extends 
functionnal pruning ideas of Rigaill 2015 and Maidstone et al. 2017 to a 
least squares criterion with a multiscale penalty promoting well-spread
changepoints. Such a penalty has been proposed in Verzelen et al. (2020), 
and it achieves optimal rates for changepoint detection and changepoint
localization (arXiv:2010.11470).
}
