% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsFPOP.R
\name{MsPELT}
\alias{MsPELT}
\title{MsPELT}
\usage{
MsPELT(y, beta, alpha, wt)
}
\arguments{
\item{y}{A vector of observations.}

\item{beta}{A constant used in the penalty.}

\item{alpha}{A constant used in the penalty.
(recommended: gamma+beta*log(length(y))).}

\item{wt}{A vector of weights associated to the observations.}
}
\value{
A list that includes the estimated position of the changepoints.
}
\description{
MsPELT is an exact segmentation algorithm that extends
inequality pruning idea from PELT to a least squares criterion with a
multiscale penalty promoting well-spread changepoints. Such a penalty 
has been proposed in Verzelen et al. (2020), and it achieves optimal rates
for changepoint detection and changepoint localization (arXiv:2010.11470).
}
